[-] Некоторые изменения

+++ libs/qsmfptr/src/tlvlist.cpp
@@ -21,18 +21,18 @@ QByteArray TLVList::to866(QString value)
 void TLVList::add(uint16_t tagId, QString tagValue)
 {
     addInt(tagId);
-    addInt(tagValue.length());
+    addInt(static_cast<uint16_t>(tagValue.length()));
     buffer.append(to866(tagValue));
 }
 
-void TLVList::add(uint16_t tagId, long tagValue, int len)
+void TLVList::add(uint16_t tagId, long tagValue, uint16_t len)
 {
     addInt(tagId);
     addInt(len);
     buffer.append(intToTLV(tagValue, len));
 }
 
-QByteArray TLVList::intToTLV(long value, uint8_t len)
+QByteArray TLVList::intToTLV(long value, uint16_t len)
 {
     if ((len > 8) || (len < 1)) {
         throw new TextException("Invalid length");
@@ -40,13 +40,13 @@ QByteArray TLVList::intToTLV(long value, uint8_t len)
     QByteArray result;
     for (int i = 0; i < len; i++)
     {
-        uint8_t b = (uint8_t) ((value >> (i * 8)) & 0xFF);
+        uint8_t b = static_cast<uint8_t>((value >> (i * 8)) & 0xFF);
         result.append(b);
     }
     return result;
 }
 
-void TLVList::addInt(int value)
+void TLVList::addInt(uint16_t value)
 {
     buffer.append((value >> 0) & 0xFF);
     buffer.append((value >> 8) & 0xFF);
