+++ libs/qsmfptr/src/tlvlist.h
@@ -8,11 +8,11 @@ class TLVList
 public:
     TLVList();
     QByteArray getData();
-    void addInt(int value);
+    void addInt(uint16_t value);
     QByteArray to866(QString value);
     void add(uint16_t tagId, QString tagValue);
-    void add(uint16_t tagId, long tagValue, int len);
-    QByteArray intToTLV(long value, uint8_t len);
+    void add(uint16_t tagId, long tagValue, uint16_t len);
+    QByteArray intToTLV(long value, uint16_t len);
 private:
     QByteArray buffer;
 };
